Moved ScoreValue to com.sap.sgs.phosphor.fosstars.model.value package
- Also moved the ScoreValueTest

This is a part of #26
