+++ src/main/java/com/sap/sgs/phosphor/fosstars/model/qa/ScoreVerifier.java
@@ -1,7 +1,7 @@
 package com.sap.sgs.phosphor.fosstars.model.qa;
 
 import com.sap.sgs.phosphor.fosstars.model.Score;
-import com.sap.sgs.phosphor.fosstars.model.ScoreValue;
+import com.sap.sgs.phosphor.fosstars.model.value.ScoreValue;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
