+++ src/main/java/com/sap/sgs/phosphor/fosstars/model/rating/AbstractRating.java
@@ -6,11 +6,11 @@ import com.sap.sgs.phosphor.fosstars.model.Label;
 import com.sap.sgs.phosphor.fosstars.model.Rating;
 import com.sap.sgs.phosphor.fosstars.model.RatingValue;
 import com.sap.sgs.phosphor.fosstars.model.Score;
-import com.sap.sgs.phosphor.fosstars.model.ScoreValue;
 import com.sap.sgs.phosphor.fosstars.model.Value;
 import com.sap.sgs.phosphor.fosstars.model.ValueSet;
 import com.sap.sgs.phosphor.fosstars.model.Version;
 import com.sap.sgs.phosphor.fosstars.model.Visitor;
+import com.sap.sgs.phosphor.fosstars.model.value.ScoreValue;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Objects;
