+++ src/test/java/com/sap/sgs/phosphor/fosstars/model/score/WeightedCompositeScoreTest.java
@@ -9,10 +9,10 @@ import static org.junit.Assert.assertTrue;
 import com.sap.sgs.phosphor.fosstars.model.Confidence;
 import com.sap.sgs.phosphor.fosstars.model.Feature;
 import com.sap.sgs.phosphor.fosstars.model.Score;
-import com.sap.sgs.phosphor.fosstars.model.ScoreValue;
 import com.sap.sgs.phosphor.fosstars.model.Value;
 import com.sap.sgs.phosphor.fosstars.model.Weight;
 import com.sap.sgs.phosphor.fosstars.model.score.WeightedCompositeScore.WeightedScore;
+import com.sap.sgs.phosphor.fosstars.model.value.ScoreValue;
 import com.sap.sgs.phosphor.fosstars.model.weight.MutableWeight;
 import java.util.Collections;
 import java.util.HashSet;
